Refactoring mostly _SET_TIMELAPSE_SETUP
Done several internal changes to increase the readability of the code.
There are also some console outputs added to the verbose option. The Hyperlapse stop message will only shown if Hyperlapse is really running, that is done as the moonraker plugin needs to send to stop request at every klipper start.