fix couple bugs arround webcam db namespace
with last couple commit the database implementation was more fragile, when there is no or faulty webcam namespace.

Now it should fallback to the default config without breaking, if someone doesn't have a 'webcams' namespace in his moonraker database (eg. client doesn't support it or it wasn't setup)

This should fix #38 and #41. Also it makes #39 obsolete.

Signed-off-by:  Christoph Frei <fryakatkop@gmail.com>